chore: revert fix NTI relationship-endpoint only returning uids DHIS2-12625
This reverts commit 0ef131fe4b4b3d70c3693ac91850cec149312ba2.

Reason being that any entity exported from NTI needs to be importable
without modification. Therefore, import/export models need to align.
