+++ dhis-2/dhis-web-api-test/src/test/java/org/hisp/dhis/webapi/controller/TrackerRelationshipsExportControllerTest.java
@@ -34,7 +34,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.Date;
 
 import org.hisp.dhis.common.IdentifiableObjectManager;
-import org.hisp.dhis.commons.util.RelationshipUtils;
 import org.hisp.dhis.jsontree.JsonArray;
 import org.hisp.dhis.jsontree.JsonObject;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
@@ -273,8 +272,9 @@ class TrackerRelationshipsExportControllerTest extends DhisControllerConvenience
         r.setFrom( rItem1 );
         r.setTo( rItem2 );
         r.setRelationshipType( type );
-        r.setKey( RelationshipUtils.generateRelationshipKey( r ) );
-        r.setInvertedKey( RelationshipUtils.generateRelationshipInvertedKey( r ) );
+        r.setKey( type.getUid() );
+        r.setInvertedKey( type.getUid() );
+        r.setAutoFields();
         manager.save( r );
         return r;
     }
@@ -289,8 +289,9 @@ class TrackerRelationshipsExportControllerTest extends DhisControllerConvenience
         r.setFrom( rItem1 );
         r.setTo( rItem2 );
         r.setRelationshipType( type );
-        r.setKey( RelationshipUtils.generateRelationshipKey( r ) );
-        r.setInvertedKey( RelationshipUtils.generateRelationshipInvertedKey( r ) );
+        r.setKey( type.getUid() );
+        r.setInvertedKey( type.getUid() );
+        r.setAutoFields();
         manager.save( r );
         return r;
     }
@@ -320,33 +321,16 @@ class TrackerRelationshipsExportControllerTest extends DhisControllerConvenience
 
     private void assertEvent( JsonObject json, ProgramStageInstance programStageInstance )
     {
-        JsonObject jsonEvent = json.getObject( "event" );
-        assertEquals( programStageInstance.getUid(), jsonEvent.getString( "event" ).string() );
-        assertEquals( programStageInstance.getStatus().toString(), jsonEvent.getString( "status" ).string() );
-        assertEquals( programStageInstance.getProgramStage().getUid(), jsonEvent.getString( "programStage" ).string() );
-        assertEquals( programStageInstance.getProgramInstance().getUid(),
-            jsonEvent.getString( "enrollment" ).string() );
-        assertTrue( jsonEvent.getArray( "relationships" ).isEmpty() );
+        assertEquals( programStageInstance.getUid(), json.getString( "event" ).string() );
     }
 
     private void assertTrackedEntity( JsonObject json, TrackedEntityInstance tei )
     {
-        JsonObject jsonTEI = json.getObject( "trackedEntity" );
-        assertEquals( tei.getUid(), jsonTEI.getString( "trackedEntity" ).string() );
-        assertEquals( tei.getTrackedEntityType().getUid(), jsonTEI.getString( "trackedEntityType" ).string() );
-        assertEquals( tei.getOrganisationUnit().getUid(), jsonTEI.getString( "orgUnit" ).string() );
-        assertTrue( jsonTEI.getArray( "relationships" ).isEmpty() );
+        assertEquals( tei.getUid(), json.getString( "trackedEntity" ).string() );
     }
 
     private void assertEnrollment( JsonObject json, ProgramInstance programInstance )
     {
-        JsonObject jsonEnrollment = json.getObject( "enrollment" );
-        assertEquals( programInstance.getUid(), jsonEnrollment.getString( "enrollment" ).string() );
-        assertEquals( programInstance.getEntityInstance().getUid(),
-            jsonEnrollment.getString( "trackedEntity" ).string() );
-        assertEquals( programInstance.getProgram().getUid(), jsonEnrollment.getString( "program" ).string() );
-        assertEquals( programInstance.getOrganisationUnit().getUid(), jsonEnrollment.getString( "orgUnit" ).string() );
-        assertTrue( jsonEnrollment.getArray( "events" ).isEmpty() );
-        assertTrue( jsonEnrollment.getArray( "relationships" ).isEmpty() );
+        assertEquals( programInstance.getUid(), json.getString( "enrollment" ).string() );
     }
 }
